+++ common/src/main/java/ru/yjailbir/commonslib/dto/request/CashRequestDtoWithToken.java
@@ -1,4 +1,4 @@
 package ru.yjailbir.commonslib.dto.request;
 
-public record CashRequestDtoWithToken(String currency, Integer value, String action, String token) {
+public record CashRequestDtoWithToken(String currency, Double value, String action, String token) {
 }
